Fix handling of hash fragment to enable a hash given to "/". (#80)
I found this while intergrating with auth0 - it redirects back to a url adding
the hash fragment. So the url is something like "/#foo=bar&baz=1". The previous
logic caused this to become just "foo=bar&baz=1" and everything goes wrong. When
we are using hash fragment routing we only want to match on "/#/" rather than
just "/#".