Update README.md
Using the function `(fn [route] (:handler route))` did not work for me. In the sample project the construction is `(comp :name :data)` which works for me with this tutorial. I also added nil handling to show a loading page since that was required for the basic application not to crash